Add a MAM endpoint for the grid crop options
